add monthly rewards breakdowns

+++ src/main/java/com/mikescarborough/demos/RewardsTracker/model/CustomerMonthRewards.java
@@ -30,11 +30,11 @@ public class CustomerMonthRewards implements java.io.Serializable {
         return rewardPoints;
     }
 
-    public float getMonthNum() {
+    public int getMonthNum() {
         return monthNum;
     }
 
-    public float getYear() {
+    public int getYear() {
         return year;
     }
